toolkit/spatial-navigation/tests/chrome/test_snav_selects.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css type="text/css"?>

<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=436084
-->

<window title="Mozilla Bug 288254"
  xmlns:html="http://www.w3.org/1999/xhtml"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
  onload="onLoad();">

  <script type="application/javascript" src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>      
  <script type="text/javascript" src="SpatialNavUtils.js"></script>

<body id="some-content" xmlns="http://www.w3.org/1999/xhtml">

<p>
<a id="start" href="https://bugzilla.mozilla.org/show_bug.cgi?id=436084">Mozilla Bug 445518</a>
</p>

<form>
<select id="select_element" name="dropdown" size="1">
  <option value ="1">option 1</option>
  <option value ="2">option 2</option>
  <option value ="3">option 3</option>
  <option value ="4">option 4</option>
</select>
</form>

<p>
<a id="end" href="https://bugzilla.mozilla.org/show_bug.cgi?id=436084">Mozilla Bug 445518</a>
</p>

</body>

<script class="testbody" type="application/javascript">
<![CDATA[
Components.utils.import("resource://gre/modules/SpatialNavigation.js");

var moveTable = [
["DOWN", "select_element"],
["UP", "start"],
["DOWN", "select_element"],
["DOWN", "select_element"],
["DOWN", "select_element"],
["DOWN", "select_element"],
["DOWN", "end"],
["DONE", "DONE"],
];

var prefs = [
["enabled",           "bool", true],
["xulContentEnabled", "bool", true],
["keyCode.modifier",  "char", "none"],
];

function onLoad()
{
    prepareTest(prefs);

    // starting the test itself.
    var x = document.getElementById("some-content");
    SpatialNavigation.init(x);
    
    // get to a known place.
    document.getElementById("start").focus();
    
    testMoves(moveTable);

    SimpleTest.waitForExplicitFinish();

    SpatialNavigation.uninit();
}

]]></script>
</window>
